+++ apps/user-portal/package.json
@@ -92,10 +92,10 @@
         "@types/lodash": "^4.14.146",
         "@types/ua-parser-js": "^0.7.33",
         "@wso2is/authenticate": "^0.1.0",
-        "@wso2is/fo2m": "^0.1.0",
+        "@wso2is/forms": "^0.1.0",
         "@wso2is/http": "^0.1.0",
         "@wso2is/theme": "^0.1.0",
-        "@wso2is/validator": "^0.1.0",
+        "@wso2is/validation": "^0.1.0",
         "history": "^4.9.0",
         "i18next": "^17.0.11",
         "i18next-browser-languagedetector": "^3.0.3",
