+++ modules/forms/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "@wso2is/fo2m",
+    "name": "@wso2is/forms",
     "version": "0.1.0",
-    "description": "Fo2m module for WSO2 Identity Server web-apps.",
+    "description": "Forms module for WSO2 Identity Server web-apps.",
     "keywords": [
         "WSO2",
         "WSO2-IS",
