+++ modules/forms/src/tests/test-form.tsx
@@ -18,7 +18,7 @@
 
 import React from "react";
 import { Form } from "semantic-ui-react";
-import { Field, Fo2m } from "..";
+import { Field, Forms } from "../../src";
 import { GroupFields } from "../components";
 import { FormField, Type, Validation } from "../models";
 import constants from "./constants";
@@ -169,7 +169,7 @@ const getForm = (testFields: FormTestFields[], isGroup?: boolean): JSX.Element =
     });
 
     return (
-        <Fo2m onSubmit={ (value) => constants.onSubmit(value) }>
+        <Forms onSubmit={ (value) => constants.onSubmit(value) }>
             {
                 isGroup
                     ? (
@@ -186,7 +186,7 @@ const getForm = (testFields: FormTestFields[], isGroup?: boolean): JSX.Element =
                         return <Field key={ index } { ...formField } />;
                     })
             }
-        </Fo2m>
+        </Forms>
     );
 
 };
