+++ modules/validation/package.json
@@ -1,12 +1,12 @@
 {
-    "name": "@wso2is/validator",
+    "name": "@wso2is/validation",
     "version": "0.1.0",
     "description": "Validator module for WSO2 Identity Server web-apps.",
     "keywords": [
         "WSO2",
         "WSO2-IS",
         "WSO2-Identity-Server",
-        "validator"
+        "validation"
     ],
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
